Add doc-comments to types missing those
Also deduplicates some types
