+++ types/src/blockchain/consensus.rs
@@ -1,18 +1,17 @@
+//! Consensus configuration, including specific consensus types like PoA
+
 use crate::{
     blockchain::primitives::BlockId,
+    fuel_crypto::{PublicKey, Signature},
     fuel_tx::Input,
     fuel_types::Address,
+    fuel_types::Bytes32,
 };
 
-mod genesis;
-mod poa;
-mod sealed;
-mod vote;
+// Different types of consensus are represented as separate modules
+pub mod poa;
 
-pub use genesis::Genesis;
-pub use poa::PoAConsensus;
-pub use sealed::Sealed;
-pub use vote::ConsensusVote;
+use poa::PoAConsensus;
 
 #[derive(Clone, Debug)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -21,6 +20,7 @@ pub use vote::ConsensusVote;
 pub enum Consensus {
     /// The genesis block defines the consensus rules for future blocks.
     Genesis(Genesis),
+    /// Proof of authority consensus
     PoA(PoAConsensus),
 }
 
@@ -45,7 +45,52 @@ impl Default for Consensus {
     }
 }
 
+/// Consensus type that a block is using
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ConsensusType {
+    /// Proof of authority
     PoA,
+}
+
+#[derive(Clone, Debug)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(any(test, feature = "test-helpers"), derive(Default))]
+/// A sealed entity with consensus info.
+pub struct Sealed<Entity> {
+    /// The actual value
+    pub entity: Entity,
+    /// Consensus info
+    pub consensus: Consensus,
+}
+
+/// A vote from a validator.
+///
+/// This is a dummy placeholder for the Vote Struct in fuel-bft
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct ConsensusVote {
+    block_id: Bytes32,
+    height: u64,
+    round: u64,
+    signature: Signature,
+    // step: Step,
+    validator: PublicKey,
+}
+
+/// The first block of the blockchain is a genesis block. It determines the initial state of the
+/// network - contracts states, contracts balances, unspent coins, and messages. It also contains
+/// the hash on the initial config of the network that defines the consensus rules for following
+/// blocks.
+#[derive(Clone, Debug, Default)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct Genesis {
+    /// The chain config define what consensus type to use, what settlement layer to use,
+    /// rules of block validity, etc.
+    pub chain_config_hash: Bytes32,
+    /// The Binary Merkle Tree root of all genesis coins.
+    pub coins_root: Bytes32,
+    /// The Binary Merkle Tree root of state, balances, contracts code hash of each contract.
+    pub contracts_root: Bytes32,
+    /// The Binary Merkle Tree root of all genesis messages.
+    pub messages_root: Bytes32,
 }
