+++ types/src/blockchain/header.rs
@@ -1,3 +1,5 @@
+//! Block header types
+
 use super::{
     consensus::ConsensusType,
     primitives::{
