--- types/src/blockchain/height.rs
@@ -1,5 +1,0 @@
-mod block_height;
-mod da_block_height;
-
-pub use block_height::BlockHeight;
-pub use da_block_height::DaBlockHeight;
