+++ types/src/entities.rs
@@ -1,2 +1,4 @@
+//! Higher level domain types
+
 pub mod coin;
 pub mod message;
