+++ types/src/entities/message.rs
@@ -1,3 +1,5 @@
+//! Message
+
 use crate::{
     blockchain::{
         header::BlockHeader,
@@ -23,17 +25,24 @@ use core::ops::Deref;
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Debug, Default, Clone, PartialEq, Eq)]
 pub struct Message {
+    /// Account that sent the message from the da layer
     pub sender: Address,
+    /// Fuel account receiving the message
     pub recipient: Address,
+    /// Nonce must be unique. It's used to prevent replay attacks
     pub nonce: Word,
+    /// The amount of the base asset of Fuel chain sent along this message
     pub amount: Word,
+    /// Arbitrary message data
     pub data: Vec<u8>,
     /// The block height from the parent da layer that originated this message
     pub da_height: DaBlockHeight,
-    pub fuel_block_spend: Option<BlockHeight>,
+    /// When the message was spent in the Fuel chain
+    pub fuel_block_spend: Option<BlockHeight>, // TODO: get rid of this
 }
 
 impl Message {
+    /// Computed message id
     pub fn id(&self) -> MessageId {
         Input::compute_message_id(
             &self.sender,
@@ -44,18 +53,21 @@ impl Message {
         )
     }
 
+    /// Compute checked message
     pub fn check(self) -> CheckedMessage {
         let id = self.id();
         CheckedMessage { message: self, id }
     }
 }
 
+/// A message associated with precomputed id
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct CheckedMessage {
     message: Message,
     id: MessageId,
 }
 
+/// Proves to da layer that this message was included in a Fuel block
 pub struct MessageProof {
     /// The proof set of the message proof.
     pub proof_set: Vec<Bytes32>,
@@ -78,6 +90,7 @@ pub struct MessageProof {
 }
 
 impl MessageProof {
+    /// Compute message id from the proof
     pub fn message_id(&self) -> MessageId {
         Output::message_id(
             &self.sender,
@@ -90,6 +103,7 @@ impl MessageProof {
 }
 
 impl CheckedMessage {
+    /// Compute message id from the proof
     pub fn id(&self) -> &MessageId {
         &self.id
     }
