+++ types/src/services.rs
@@ -1,1 +1,3 @@
+//! Types for specific services
+
 pub mod txpool;
