+++ types/src/services/txpool.rs
@@ -1,3 +1,5 @@
+//! Types for interoperability with the txpool service
+
 use crate::{
     blockchain::primitives::BlockId,
     fuel_asm::Word,
@@ -30,13 +32,16 @@ use std::{
 };
 use tai64::Tai64;
 
+/// Pool transaction wrapped in an Arc for thread-safe sharing
 pub type ArcPoolTx = Arc<PoolTransaction>;
 
 /// Transaction type used by the transaction pool. Transaction pool supports not
 /// all `fuel_tx::Transaction` variants.
 #[derive(Debug, Eq, PartialEq)]
 pub enum PoolTransaction {
+    /// Script
     Script(Checked<Script>),
+    /// Create
     Create(Checked<Create>),
 }
 
@@ -72,6 +77,7 @@ impl UniqueIdentifier for PoolTransaction {
     }
 }
 
+#[allow(missing_docs)]
 impl PoolTransaction {
     pub fn is_computed(&self) -> bool {
         match self {
@@ -142,16 +148,20 @@ impl From<Checked<Create>> for PoolTransaction {
 /// of the `inserted` transaction.
 #[derive(Debug)]
 pub struct InsertionResult {
+    /// This was inserted
     pub inserted: ArcPoolTx,
+    /// These were removed during the insertion
     pub removed: Vec<ArcPoolTx>,
 }
 
+/// Information of a transaction fetched from the txpool
 #[derive(Debug, Clone)]
 pub struct TxInfo {
     tx: ArcPoolTx,
     submitted_time: Tai64,
 }
 
+#[allow(missing_docs)]
 impl TxInfo {
     pub fn new(tx: ArcPoolTx) -> Self {
         Self {
@@ -180,26 +190,40 @@ impl Deref for TxInfo {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub enum TransactionStatus {
+    /// Transaction was submitted into the txpool
     Submitted {
+        /// Timestamp of submission into the txpool
         time: Tai64,
     },
+    /// Transaction was successfully included in a block
     Success {
+        /// Included in this block
         block_id: BlockId,
+        /// Time when the block was generated
         time: Tai64,
+        /// Result of executing the transaction for scripts
         result: Option<ProgramState>,
     },
+    /// Transaction was squeezed of the txpool
     SqueezedOut {
+        /// Why this happened
         reason: String,
     },
+    /// Transaction was included in a block, but the exection was reverted
     Failed {
+        /// Included in this block
         block_id: BlockId,
+        /// Time when the block was generated
         time: Tai64,
+        /// Why this happened
         reason: String,
+        /// Result of executing the transaction for scripts
         result: Option<ProgramState>,
     },
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
+#[allow(missing_docs)] // TODO: This type is likely removed soon
 pub enum TxStatus {
     /// Submitted into txpool.
     Submitted,
